+++ Terra/src/utopia/terra/controller/coordinate/world/WorldView.scala
@@ -1,5 +1,8 @@
 package utopia.terra.controller.coordinate.world
 
+import utopia.paradigm.measurement.Distance
+import utopia.terra.model.world.WorldDistance
+
 /**
  * Common trait for different world representation models used for coordinate transformations
  * @author Mikko Hilpinen
@@ -13,3 +16,15 @@ trait WorldView[V2D, -V3D, +Surface, +Aerial]
 	extends LatLongToWorldPoint[Surface, Aerial] with VectorToWorldPoint[V2D, V3D, Surface, Aerial]
 		with VectorDistanceConversion with LatLongFromVectorFactory[V2D]
 		with VectorFromLatLongFactory[V2D]
+{
+	/**
+	  * @param distance A distance travelled
+	  * @return Copy of that distance as a "world distance"
+	  */
+	def distance(distance: Distance): WorldDistance = WorldDistance(distance)(this)
+	/**
+	  * @param vectorDistance A distance travelled (in vector space)
+	  * @return Copy of that distance as a "world distance"
+	  */
+	def distance(vectorDistance: Double): WorldDistance = WorldDistance.vector(vectorDistance)(this)
+}