+++ Terra/src/utopia/terra/model/world/grid/GridPointOps.scala
@@ -3,7 +3,7 @@ package utopia.terra.model.world.grid
 import utopia.flow.operator.EqualsBy
 import utopia.paradigm.measurement.Distance
 import utopia.paradigm.shape.template.HasDimensions.HasDoubleDimensions
-import utopia.paradigm.shape.template.vector.DoubleVectorLike
+import utopia.paradigm.shape.template.vector.{DoubleVector, DoubleVectorLike}
 import utopia.terra.controller.coordinate.world.GridArea
 import utopia.terra.model.CompassTravel
 import utopia.terra.model.angular.{CompassRotation, LatLong, LatLongRotation}
@@ -15,14 +15,14 @@ import utopia.terra.model.world.{WorldPoint, WorldPointOps}
   * @since 2.11.2023, v1.0.1
   */
 trait GridPointOps[+V <: DoubleVectorLike[V], +Repr]
-	extends WorldPointOps[V, WorldPoint[HasDoubleDimensions], Repr] with EqualsBy
+	extends WorldPointOps[V, WorldPoint[HasDoubleDimensions], DoubleVector, Repr] with EqualsBy
 {
 	// ABSTRACT ----------------------
 	
 	/**
 	  * @return The grid system used to transform coordinates
 	  */
-	protected implicit def grid: GridArea
+	implicit def grid: GridArea
 	
 	/**
 	  * @param surfaceLocation The new surface location to assign
@@ -48,4 +48,5 @@ trait GridPointOps[+V <: DoubleVectorLike[V], +Repr]
 	override def +(travel: CompassTravel): Repr = at(vector + travel.axis(grid.vectorLengthOf(travel.distance)))
 	override def +(rotation: CompassRotation): Repr = at(latLong + rotation)
 	override def +(other: LatLongRotation): Repr = at(latLong + other)
+	override def +(vectorTravel: DoubleVector): Repr = at(vector + vectorTravel)
 }