+++ Paradigm/Changes.md
@@ -11,6 +11,8 @@
 ### New methods
 - **Bounds**
   - Added `.relativize(...)` and `.relativeToAbsolute(...)` methods that perform **Bounds**-based coordinate conversions
+- **Distance** (object)
+  - Added `.ofKilometers(Double)`
 
 ## v1.4 - 27.09.2023
 This version applies major under-the-hood refactoring to vector classes,